backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
supabase
fpdf2
pandas
python-multipart
